refactor(imageLoader): move isColorImage to shared directory
This commit moves the isColorImage function from the imageLoader directory to the shared directory to be used across the dicomImageLoader package. It also updates the imports of isColorImage throughout the dicomImageLoader package. The isColorImage function checks whether the image is a color image or not based on the photometric interpretation. The commit also updates the createImage and decodeImageFrame functions to use the isColorImage function and render color image in 8-bit in case of rendering issues.
