+++ packages/dicomImageLoader/src/shared/isColorImage.ts
